fix: add patches tab
